Beginning of my alarm clock project. Contains 3 projects
menu_system: The menu system I'm building to set time and 7 alarms on
the alarm clock.
alarm_clock: The state machine behind the alarm clock.
These two will eventually be merged into a single alarm clock project
(and will gain the ability to store alarm times in the EEPROM)

serlcd_charset: This was a demo to acquaint myself with my SerLCD.