If the menu has been idle for 1 minute, cancel it.
The toddler started playing with my alarm clock over a 3 day Yom Tov,
and left it in menu mode until I could push more buttons after Shabbat.
This protection ensures that if he plays with it again on Shabbat, it
goes back to displaying the time. (Also, the menu color is too bright
and is annoying when we're trying to  sleep, so we really want it to go
back to displaying red.)