When printing the main display, always fill every cell.
This adds padding to the writes in PrintTimeTall, so that every time
through the display loop we write to every cell in the display. This
overwrites the garbage characters that occasionally appear on the
display. I don't know why this happens, but it improved a lot when I cut
some of the I2C pull-up resistors, and then it reappeared when the
weather got cold and dry.